+++ docs/CHANGELOG.md
@@ -59,6 +59,7 @@ Version 1.0.0
 - Stop mover brushes being split to "fix PS2 tiling" in editor
 - Add framework for real-time modification of defined event properties
 - Add custom events `SetVar`, `Difficulty_Gate`, `HUD_Message`, `Play_Video`, `Set_Level_Hardness`, `Sequence`, `Clear_Queued`, `Remove_Link`, `Fixed_Delay`, `Add_Link`, `Valid_Gate`
+- Add goal (variable) logic events `Goal_Math` and `Goal_Gate`